tests/requirements.txt
asyncio
csvkit
Flask
matplotlib
metaapi-cloud-sdk
numpy==1.24.4
pandas
pandas-ta
scikit-learn==1.3.2
gunicorn
Flask
plotly
TextBlob
yfinance
python-docx